Refactor format code into its own package
