+++ toolkit/tools/pkg/licensecheck/licensecheck.go
@@ -137,79 +137,24 @@ func (l *LicenseChecker) CheckLicenses() error {
 	return nil
 }
 
-// GetAllResults returns the results of the search, split into:
+// GetResults returns the results of the search, split into:
 // - All results: Every scan result
 // - Any result that has at least one warning
 // - Any result that has at least one error
-// There may be overlap between the lists. If pedantic is true, warnings will be treated as errors.
-func (l *LicenseChecker) GetAllResults(pedantic bool) (all, warnings, errors []LicenseCheckResult) {
+func (l *LicenseChecker) GetResults() (all, warnings, errors []LicenseCheckResult) {
 	all = []LicenseCheckResult{}
 	warnings = []LicenseCheckResult{}
 	errors = []LicenseCheckResult{}
 	for _, result := range l.results {
 		all = append(all, result)
-		if result.HasBadResult() || (pedantic && result.HasWarningResult()) {
+		if result.HasBadResult() {
 			errors = append(errors, result)
 		}
-		if result.HasWarningResult() && !pedantic {
+		if result.HasWarningResult() {
 			warnings = append(warnings, result)
 		}
 	}
 	return all, warnings, errors
-}
-
-// FormatResults formats the results of the search to a string. Results will be ordered as follows:
-// - Packages with warnings only, sorted alphabetically
-// - Packages with errors (and possibly warnings), sorted alphabetically
-// If pedantic is true, warnings will be treated as errors.
-func (l *LicenseChecker) FormatResults(pedantic bool) string {
-	var sb strings.Builder
-	results := sortAndFilterResults(l.results)
-
-	// Print warnings first, but only if they don't also have an error
-	if !pedantic {
-		for _, result := range results {
-			if result.HasWarningResult() && !result.HasBadResult() {
-				sb.WriteString(printWarning(result))
-			}
-		}
-	}
-
-	// Now print the errors
-	for _, result := range results {
-		if result.HasBadResult() || (pedantic && result.HasWarningResult()) {
-			sb.WriteString(printError(result, pedantic))
-			if !pedantic && result.HasWarningResult() {
-				// If pedantic was set, the warning was already printed as an error.
-				// Otherwise print the warning now.
-				sb.WriteString(printWarning(result))
-			}
-		}
-	}
-
-	return sb.String()
-}
-
-func printWarning(result LicenseCheckResult) string {
-	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("WARN: '%s' has license warnings:\n", filepath.Base(result.RpmPath)))
-	sb.WriteString(fmt.Sprintf("\tduplicated license files:\n\t\t%s\n", strings.Join(result.DuplicatedDocs, "\n\t\t")))
-	return sb.String()
-}
-
-func printError(result LicenseCheckResult, pedantic bool) string {
-	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("ERROR: '%s' has license errors:\n", filepath.Base(result.RpmPath)))
-	if len(result.BadDocs) > 0 {
-		sb.WriteString(fmt.Sprintf("\tbad %%doc files:\n\t\t%s\n", strings.Join(result.BadDocs, "\n\t\t")))
-	}
-	if len(result.BadFiles) > 0 {
-		sb.WriteString(fmt.Sprintf("\tbad general file:\n\t\t%s\n", strings.Join(result.BadFiles, "\n\t\t")))
-	}
-	if pedantic && len(result.DuplicatedDocs) > 0 {
-		sb.WriteString(fmt.Sprintf("\tduplicated license files:\n\t\t%s\n", strings.Join(result.DuplicatedDocs, "\n\t\t")))
-	}
-	return sb.String()
 }
 
 // runLicenseCheckInChroot searches for bad licenses amongst the RPMs mounted into the chroot. This function is meant
