+++ toolkit/tools/pkg/licensecheck/licensecheck_test.go
@@ -105,163 +105,6 @@ func TestIsFileMisplacedInLicensesFolderDetectPackageFolder(t *testing.T) {
 	assert.True(t, isFileMisplacedInLicensesFolder("/usr/share/licenses/OTHER_PKG", emptyLicenseFiles))
 }
 
-func TestFormatResultsNonPedantic(t *testing.T) {
-	const notPedantic = false
-	testCases := []struct {
-		name     string
-		results  []LicenseCheckResult
-		expected string
-	}{
-		{
-			name:     "No results",
-			results:  []LicenseCheckResult{},
-			expected: "",
-		},
-		{
-			name: "Single result",
-			results: []LicenseCheckResult{
-				{
-					RpmPath: "/path/to/package.rpm",
-					BadDocs: []string{"doc1"},
-				},
-			},
-			expected: "ERROR: 'package.rpm' has license errors:\n" +
-				"\tbad %doc files:\n" +
-				"\t\tdoc1\n",
-		},
-		{
-			name: "Multiple results",
-			results: []LicenseCheckResult{
-				{
-					RpmPath:        "/path/to/package.rpm",
-					BadDocs:        []string{"doc1", "doc2"},
-					BadFiles:       []string{"file1", "file2"},
-					DuplicatedDocs: []string{"dupe1", "dupe2"},
-				},
-				{
-					RpmPath:        "/path/to/another-package.rpm",
-					DuplicatedDocs: []string{"dupe3", "dupe4"},
-				},
-			},
-			expected: "WARN: 'another-package.rpm' has license warnings:\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe3\n" +
-				"\t\tdupe4\n" +
-				"ERROR: 'package.rpm' has license errors:\n" +
-				"\tbad %doc files:\n" +
-				"\t\tdoc1\n" +
-				"\t\tdoc2\n" +
-				"\tbad general file:\n" +
-				"\t\tfile1\n" +
-				"\t\tfile2\n" +
-				"WARN: 'package.rpm' has license warnings:\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe1\n" +
-				"\t\tdupe2\n",
-		},
-		{
-			name: "Duplicated docs only",
-			results: []LicenseCheckResult{
-				{
-					RpmPath:        "/path/to/package.rpm",
-					DuplicatedDocs: []string{"dupe1", "dupe2"},
-				},
-			},
-			expected: "WARN: 'package.rpm' has license warnings:\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe1\n" +
-				"\t\tdupe2\n",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			checker := LicenseChecker{
-				results: tc.results,
-			}
-			actual := checker.FormatResults(notPedantic)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
-func TestFormatResultsPedantic(t *testing.T) {
-	const pedantic = true
-	testCases := []struct {
-		name     string
-		results  []LicenseCheckResult
-		expected string
-	}{
-		{
-			name:     "No results",
-			results:  []LicenseCheckResult{},
-			expected: "",
-		},
-		{
-			name: "Single result",
-			results: []LicenseCheckResult{
-				{
-					RpmPath: "/path/to/package.rpm",
-					BadDocs: []string{"doc1"},
-				},
-			},
-			expected: "ERROR: 'package.rpm' has license errors:\n" +
-				"\tbad %doc files:\n" +
-				"\t\tdoc1\n",
-		},
-		{
-			name: "Multiple results",
-			results: []LicenseCheckResult{
-				{
-					RpmPath:        "/path/to/package.rpm",
-					BadDocs:        []string{"doc1", "doc2"},
-					BadFiles:       []string{"file1", "file2"},
-					DuplicatedDocs: []string{"dupe1", "dupe2"},
-				},
-				{
-					RpmPath:        "/path/to/another-package.rpm",
-					DuplicatedDocs: []string{"dupe3", "dupe4"},
-				},
-			},
-			expected: "ERROR: 'another-package.rpm' has license errors:\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe3\n" +
-				"\t\tdupe4\n" +
-				"ERROR: 'package.rpm' has license errors:\n" +
-				"\tbad %doc files:\n" +
-				"\t\tdoc1\n" +
-				"\t\tdoc2\n" +
-				"\tbad general file:\n" +
-				"\t\tfile1\n" +
-				"\t\tfile2\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe1\n" +
-				"\t\tdupe2\n",
-		},
-		{
-			name: "Duplicated docs only",
-			results: []LicenseCheckResult{
-				{
-					RpmPath:        "/path/to/package.rpm",
-					DuplicatedDocs: []string{"dupe1", "dupe2"},
-				},
-			},
-			expected: "ERROR: 'package.rpm' has license errors:\n" +
-				"\tduplicated license files:\n" +
-				"\t\tdupe1\n" +
-				"\t\tdupe2\n",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			checker := LicenseChecker{
-				results: tc.results,
-			}
-			actual := checker.FormatResults(pedantic)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
 func makeResult(name string, numBadDocs, numBadFiles, numDupes int) LicenseCheckResult {
 	badDocs := make([]string, numBadDocs)
 	for i := 0; i < numBadDocs; i++ {
@@ -393,7 +236,7 @@ func TestGetResults(t *testing.T) {
 			checker := LicenseChecker{
 				results: tc.results,
 			}
-			all, warn, fail := checker.GetAllResults(false)
+			all, warn, fail := checker.GetResults()
 			assert.Equal(t, tc.expected.all, all)
 			assert.Equal(t, tc.expected.warn, warn)
 			assert.Equal(t, tc.expected.fail, fail)
